requirements.txt
Flask==3.0.3
fuzzywuzzy==0.18.0
openai==1.37.1
python-dotenv==1.0.1
